pyproject.toml
[tool.poetry]
name = "zhenxun_bot"
version = "0.1.1"
description = "基于 Nonebot2 和 go-cqhttp 开发，以 postgresql 作为数据库，非常可爱的绪山灵梦bot"
authors = ["HibiKier <775757368@qq.com>"]
license = "AGPL"

[[tool.poetry.source]]
name = "ali"
default = true
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-beta.2"
nonebot-adapter-onebot = "^2.0.0-beta.1"
aiofiles = "^0.8.0"
aiohttp = "3.7.4.post0"
beautifulsoup4 = "4.9.3"
feedparser = "^6.0.8"
gino = "^1.0.1"
httpx = "^0.22.0"
ImageHash = "^4.2.1"
jieba = "^0.42.1"
lxml = "4.6.5"
nonebot-plugin-apscheduler = "^0.1.2"
opencv-python = "^4.5.5"
Pillow = "^9.0.1"
playwright = "^1.18.2"
psutil = "^5.9.0"
PyYAML = "5.4.1"
retrying = "^1.3.3"
ujson = "^5.1.0"
"ruamel.yaml" = "^0.17.21"
matplotlib = "^3.5.1"
black = "^22.1.0"
pypinyin = "^0.46.0"
dateparser = "^1.1.0"
cn2an = "^0.5.16"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
bilireq = "^0.1.2"
emoji = "^1.7.0"
wordcloud = "^1.8.1"
rich = "^12.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
